Added new data
Today I added new data for the second time this weekend (forgot to do it on Friday). Unfortunately, this code has been a little buggy and will not display all of the markers. I think that this may have something to do with my efforts at adding a feature to display different types of data. Those changes are not yet functioning, so are not in this build. Until next time!